# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "The user service handles the whole user management. On start the service will check if the administrator account exist, if not, it will create and adminstration user with the given configuration.",
    "version": "1.0.0",
    "title": "2ndLock user service",
    "termsOfService": "http://2ndLock.com",
    "contact": {
      "email": "jens@symlink.de"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/",
...
  "schemes": ["http"],
  "paths": {
...
    "/password/forgot": {
      "post": {
        "tags": ["password"],
        "summary": "send password forgot request",
        "responses": {
          "200": {
            "description": "success"
          },
          "400": {
            "description": "bad request"
          }
        }
      }
    },
    "/password/update": {
      "post": {
        "tags": ["password"],
        "summary": "set new password for account",
        "responses": {
          "200": {
            "description": "success"
          },
          "400": {
            "description": "bad request"
          }
        }
      }
    },
...
  }
}